requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==1.4.49
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
stripe==2.67.0
python-dotenv==0.20.0
Pillow==9.2.0
qrcode==7.3.1
requests==2.28.1